Fixed GCC compatability of scripting module:
* CMake build files now support unix builds
* SWIG interface files and Lua wrappers are now fully C++ compliant to comfort GCC
* Hacked shared/static builds of the campvis-modules module even more, but now it works on both MSVC and GCC

+++ core/tools/concurrenthistogram.h
@@ -164,7 +164,7 @@ namespace campvis {
     template<typename T, size_t ND>
     void campvis::ConcurrentGenericHistogramND<T, ND>::addSample(T sample[ND]) {
         size_t bucketNumbers[ND];
-        for(int i = 0; i < ND; ++i)
+        for (size_t i = 0; i < ND; ++i)
             bucketNumbers[i] = getBucketNumber(i, sample[i]);
 
         size_t index = getArrayIndex(bucketNumbers);
